runtime/riscv.S: add DWARF CFI directives
This is needed for backtraces to function in GDB.

CFI_{START,END}PROC is added to FUNCTION/END_FUNCTION.
.Ltrap_handler needs its own CFI_STARTPROC because it doesn't use
FUNCTION.

The rules for CFI_ADJUST are fairly simple: whenever the `sp` is
manipulated add a CFI_ADJUST right after it with an opposite sign.
This allows the unwinder to always find the frame even when the sp
gets moved multiple times.

CFI_OFFSET is used to tell the unwinder where the various registers are
saved on the stack. At a minimum we need to tell it where to find `ra`,
so that it can continue unwinding the stack.
The value for CFI_OFFSET is calculated by undoing the adjustment
made by CFI_ADJUST (so we get the value of the sp from the frame),
and then adding the same offset use by the STORE instruction.

Signed-off-by: Edwin Török <edwin@etorok.net>
(cherry picked from commit d4b73d36a31670ef5c1fb1d707789cab02d2d2ab)