asmcomp/riscv: emit CFI directives
Needed for stacktraces to work in a debugger.
cfi_startproc/cfi_endproc marks the boundaries of a function.

adjust_stack_offset updates the offset between the stackpointer and the
frame as the stackpointer is adjusted the offset is adjusted in opposite
direction.
Every `addi sp` needs a corresponding adjust_cfa_offset.
This is not done for tailcalls where other backends have a comment about
undoing CFA adjustments because the function may continue.

cfi_offset is for telling the unwinder where to find register on the
stack, at a minimum we need to tell it about the location of `ra`.

Note that there is a difference here: in OCaml code it is at c-8,
whereas in caml_call_gc it is at c-16
(c-8 is unused gap due to alignment).

This change was modeled after arm64/emit.mlp

Signed-off-by: Edwin Török <edwin@etorok.net>
(cherry picked from commit e015077bf0cddd1fb90383f2b7fb948177cad990)