ocamlopt: add -Wl,--gc-sections when -nodynlink is used
`-ffunction-sections` places each function into its own
`.text.caml.<name>` section.
`-Wl,--gc-sections` removes the sections that are not referenced by the
main (`.text`, `.data`, ...) sections.

When using the `Dynlink` module dynamically loaded libraries may need to
refer back to symbols in the main executable, so only add this flag when
`-nodynlink` is used.

To see the list of removed sections `-ccopt -Wl,--print-gc-sections` can
be used, and to see why a particular symbol is not removed: `-ccopt
-Wl,--trace-symbol=<symbol-name>`.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>