use .data.caml.dynlink  for primitive references
The section is marked as mergable, so that any reference to a symbol in
the section keeps all the symbols alive.
By default we'll generate a single reference from `.data` to
`.data.caml.dynlink` to keep this section (and the symbols they
reference) alive.

When `-nodynlink` is used then we do not generate this reference,
which allows the linker to drop all the primitives from the executable.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>