Sanitizers: split into individual scripts
This enables developers to debug issues raised by a single sanitizer
more easily by testing just that sanitizer.

'git clean' is only called in the toplevel 'CI' script, so that
the sanitizer scripts can be called with uncommitted changes too.

This split will be useful for documenting how to use sanitizers
in runtime/HACKING.adoc.

Recommended to review with these git settings that shows
lines that get moved unchanged in a different color:
```
[diff]
    algorithm = histogram
    colorMoved = dimmed_zebra
```

No functional change.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>